refactor: split plugins.ts into focused modules under plugins/
Split the 1,094-line plugins.ts monolith into 5 focused modules inside
a new src/plugins/ subdirectory:

  plugins/types.ts     Constants, readFileCapped utility, all TypeScript interfaces
  plugins/parse.ts     Frontmatter, manifest, command, hook, skill, agent, rule parsers
  plugins/discover.ts  Directory discovery and Claude Code marketplace integration
  plugins/hooks.ts     Hook event mapping (Claude → pi) and tool matcher logic
  plugins/info.ts      Lightweight PluginInfo serialization for the Web UI / API
  plugins/index.ts     Barrel re-export of all 5 modules

plugins.ts is now a 1-line re-export: export * from './plugins/index.js'
All existing callers are unchanged. No logic changes.
145 tests pass, typecheck clean.
